changes employee/balance INTEGER

+++ src/main/java/com/open3hr/adeies/entities/LeaveBalance.java
@@ -17,13 +17,13 @@ public class LeaveBalance {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "days")
-    private Long days;
+    private Integer days;
 
     @Column(name = "days_taken")
-    private Long daysTaken;
+    private Integer daysTaken;
 
     @OneToMany(mappedBy = "",
             fetch = FetchType.EAGER,
@@ -37,7 +37,7 @@ public class LeaveBalance {
                     CascadeType.REFRESH,
             })
     @JoinColumn(name = "leave_category_id", referencedColumnName = "id")
-    private Long leaveCategoryId;
+    private Integer leaveCategoryId;
 
     public LeaveBalance(LeaveBalanceDTO leaveBalanceDTO){
         this.id = leaveBalanceDTO.getId();
